// src/app/home/models/movieAPIdetails.model.ts
export interface StrukturaModel {
    Source: string;
    Value: string;
}
export interface MovieAPIdetails {
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;

    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;

    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Ratings: StrukturaModel[];
    // Ratings: [{Source: string, Value: string}];

    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;

    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
    Response: string;
}